On branch viewsController Refactor: Created a dedicated controller for returning views Changes to be committed: 	modified:   src/controllers/mongo/carts.controller.js 	modified:   src/controllers/mongo/products.controller.js 	modified:   src/controllers/mongo/users.controller.js 	new file:   src/controllers/views.controller.js 	modified:   src/routers/views/carts.view.router.js 	modified:   src/routers/views/products.view.api.js 	modified:   src/routers/views/users.view.api.js
