+++ src/controllers/mongo/carts.controller.js
@@ -5,7 +5,6 @@ class CartsController extends MongoCrudController {
   constructor() {
     super(cartsService, "cart");
     this.read = this.read.bind(this);
-    this.showCart = this.showCart.bind(this);
   }
 
   async read(req, res) {
@@ -15,16 +14,6 @@ class CartsController extends MongoCrudController {
       res.json200(instance, `fetched ${this.modelName} from user with id ${id}`) 
       :
       res.json404({}, `${this.modelName} for user with id ${id} was not found`)
-  }
-
-  async showCart(req, res, next) {
-    try {
-      const { id } = req.params;
-      const cart = await cartsService.read(id);
-      return res.render("cart.handlebars", { cart });
-    } catch (error) {
-      throw error;
-    }
   }
 }
