+++ src/routers/views/users.view.api.js
@@ -1,13 +1,13 @@
 import CustomRouter from "../customRouter.js";
-import usersController from "../../controllers/mongo/users.controller.js";
+import viewsController from "../../controllers/views.controller.js";
 
 class UsersRouter extends CustomRouter {
   constructor() {
     super();
-    this.read("/login", ["PUBLIC"],usersController.showLogin);
-    this.read("/register",["PUBLIC"], usersController.showRegister);
-    this.read("/verify/:id",["PUBLIC"], usersController.showVerify);
-    this.read("/:userId", ["PUBLIC"], usersController.showUser);
+    this.read("/login", ["PUBLIC"],viewsController.showLogin);
+    this.read("/register",["PUBLIC"], viewsController.showRegister);
+    this.read("/verify/:id",["PUBLIC"], viewsController.showVerify);
+    this.read("/:userId", ["PUBLIC"], viewsController.showUser);
   }
 }
